+++ tests/main.c
@@ -4,17 +4,17 @@
 #include "../src/grouped_table.h"
 
 
-declareVector(IntVector, int, intVector)
-defineVectorStruct(IntVector, int, intVector)
-defineVector(IntVector, int, intVector)
+declareVector(IntVector, int16_t, intVector)
+defineVectorStruct(IntVector, int16_t, intVector)
+defineVector(IntVector, int16_t, intVector)
 
 
-declareGroupedTable(IntByIntTable, int, int, IntVector, intByIntTable)
+declareGroupedTable(IntByIntTable, int16_t, int16_t, IntVector, intByIntTable)
 declareVector(IntByIntTableRowsVector, IntByIntTableRow, intByIntTableRowsVector)
 defineVectorStruct(IntByIntTableRowsVector, IntByIntTableRow, intByIntTableRowsVector)
 defineVector(IntByIntTableRowsVector, IntByIntTableRow, intByIntTableRowsVector)
 #define intsEqual(a, b) a == b
-defineGroupedTable(IntByIntTable, int, int, IntVector, intVector, intByIntTable, intsEqual)
+defineGroupedTable(IntByIntTable, int16_t, int16_t, IntVector, intVector, intByIntTable, intsEqual)
 
 
 void test_setup(void) {
@@ -26,19 +26,19 @@ void test_teardown(void) {
 
 
 MU_TEST(test_vector) {
-	int capacity = 10;
+	int16_t capacity = 10;
 	IntVector vector = makeIntVector(capacity);
 
 	{
-		int i;
+		int16_t i;
 		for (i=0; i<capacity; ++i) {
 			intVectorPush(vector, i);
 		}
 	}
 
 	{
-		int *element;
-		int i;
+		int16_t *element;
+		int16_t i;
 		vectorForeachIndex(intVectorBegin(vector), intVectorEnd(vector), element, i) {
 			// printf("%i => %i\n", i, *element);
 			mu_assert_int_eq(i, *element);
@@ -50,7 +50,7 @@ MU_TEST(test_vector) {
 
 
 MU_TEST(test_coods_grouped_by_strip) {
-	int capacity = 10;
+	int16_t capacity = 10;
 	IntByIntTable table = makeIntByIntTable(capacity);
 
 	intByIntTableAdd(table, 1, 1);
@@ -67,7 +67,7 @@ MU_TEST(test_coods_grouped_by_strip) {
 
 	{
 		IntByIntTableRow *row;
-		int i;
+		int16_t i;
 		vectorForeachIndex(intByIntTableBegin(table), intByIntTableEnd(table), row, i) {
 			// printf("%i => %i\n", row->key, intVectorSize(row->values));
 			if (i == 0) {
@@ -92,7 +92,7 @@ MU_TEST_SUITE(test_suite) {
 }
 
 
-int main(int argc, char *argv[]) {
+int16_t main(int16_t argc, char *argv[]) {
 	MU_RUN_SUITE(test_suite);
 	MU_REPORT();
 	return MU_EXIT_CODE;
